:bug: Fix Input.Password cannot get input element
close #18412
